Create RobustSynchronizationModule.ino
Expanding the AdvancedSynchronizationModule.ino to include error correction and Automatic Repeat reQuest (ARQ), we can add CRC error checking to the packets and retransmission requests if CRC doesn't match. I am also adding a basic retransmission mechanism.

This updated example includes CRC checking for error detection and an ARQ mechanism for requesting retransmission of packets. Note that this code is still a conceptual example and the actual communication logic for sending, receiving and acknowledgment of packets need to be implemented based on the RF module used.