Create CompleteSynchronizationModule.ino
In the code below, I've expanded the RobustSynchronizationModule.ino to incorporate communication logic using the ESP32 as the RF module communicating with the SAMD51 via SPI.

For communication, I'm assuming that the ESP32 is loaded with firmware that allows it to function as a packet radio transceiver and is capable of sending and receiving packets over the 2.4GHz band.

Please note that the code is still in an abstract form as the specific SPI command set and protocol for communication with the ESP32 will depend on the firmware loaded on the ESP32. This example is more like a skeleton that you can fill in with the specific command set required for your particular application.

Please remember that this is still a skeleton example. The specific commands and protocol to interact with the ESP32 over SPI would need to be implemented based on the ESP32's firmware and documentation.