Create EncryptionModule.ino
Below is an example of an Arduino code for an Encryption Module (EncryptionModule.ino) that can be used in a SAMD51-based system. This module is responsible for encrypting and decrypting data that will be sent or has been received via the FHSS communication.

In this example, I use the AES (Advanced Encryption Standard) block cipher for encrypting and decrypting the data.

This example code uses the AESLib library, which is an AES encryption library for Arduino. You will need to add the AESLib library to your Arduino IDE before you can compile the code. You can install it through the Arduino Library Manager or download it from the GitHub repository and add it manually to your Arduino libraries.

Please note that in a real-world application, the encryption keys should be handled securely. This example serves as a basic illustration of how data can be encrypted and decrypted using AES encryption. Integrating this code with the actual communication and key exchange mechanisms should be done carefully, taking into account best practices for secure communication.