Create DeviceAuthenticationModule.ino
Below is an example code for a Device Authentication Module (DeviceAuthenticationModule.ino) that can be used in a SAMD51-based system. In this example, I'm using HMAC (Hash-based Message Authentication Code) with the SHA-256 hashing algorithm as an authentication method to ensure that the data sent between devices is from an authenticated source and has not been tampered with.

Please note that security and cryptography are complex subjects and this example is meant to be illustrative. In a production environment, it's advisable to use well-vetted libraries and consult with security experts.

This example code uses the Crypto library, which provides cryptographic functions for Arduino. You will need to add the Crypto library to your Arduino IDE before you can compile this code. You can install it through the Arduino Library Manager or download it from GitHub and add it manually to your Arduino libraries.

This module can be integrated with the rest of your Frequency-hopping spread spectrum (FHSS) system to ensure that only authenticated devices can communicate with each other and that the integrity of the messages is maintained.