Create DynamicInverseDemultiplexerModule.ino
For the receiver side, an inverse demultiplexer module is necessary to reassemble the split data streams into the original high-data-rate stream. To achieve this, the receiver must be able to synchronize with the transmitter, receive data on multiple channels, and reassemble the data in the correct order.

In this code, the receiveAndReassembleData function is responsible for receiving data chunks from multiple channels and reassembling them into the original data stream. Note that this code assumes that the data chunks can be reassembled in the order they are received, which may not always